requirements.txt
docker-compose>=1.23.1
envparse>=0.2.0
pika>=0.12.0
psycopg2-binary>=2.7.6.1
